+++ templates/slots/Services/COPage/PageComponent/classes/class.il__PLUGIN_NAME__PluginGUI.php
@@ -5,8 +5,7 @@ use srag\DIC\__PLUGIN_NAME__\DICTrait;
 /**
  * Class il__PLUGIN_NAME__PluginGUI__VERSION_COMMENT__
  *
- * @author            studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
- * @author            __RESPONSIBLE_NAME__ <__RESPONSIBLE_EMAIL__>
+ * __AUTHOR_COMMENT__
  *
  * @ilCtrl_isCalledBy il__PLUGIN_NAME__PluginGUI: ilPCPluggedGUI
  */
@@ -17,7 +16,6 @@ class il__PLUGIN_NAME__PluginGUI extends ilPageComponentPluginGUI
     const PLUGIN_CLASS_NAME = il__PLUGIN_NAME__Plugin::class;
     const CMD_CANCEL = "cancel";
     const CMD_CREATE = "create";
-    const CMD_CREATE_PLUG = "create_plug";
     const CMD_EDIT = "edit";
     const CMD_INSERT = "insert";
     const CMD_UPDATE = "update";
@@ -33,9 +31,9 @@ class il__PLUGIN_NAME__PluginGUI extends ilPageComponentPluginGUI
 
 
     /**
-     *
+     * @inheritDoc
      */
-    public function executeCommand()
+    public function executeCommand()/*:void*/
     {
         $next_class = self::dic()->ctrl()->getNextClass($this);
 
@@ -75,18 +73,18 @@ class il__PLUGIN_NAME__PluginGUI extends ilPageComponentPluginGUI
 
 
     /**
-     *
+     * @inheritDoc
      */
-    public function insert()
+    public function insert()/*:void*/
     {
         $this->edit();
     }
 
 
     /**
-     *
+     * @inheritDoc
      */
-    public function create()
+    public function create()/*:void*/
     {
         $form = $this->getForm();
 
@@ -110,9 +108,9 @@ class il__PLUGIN_NAME__PluginGUI extends ilPageComponentPluginGUI
 
 
     /**
-     *
+     * @inheritDoc
      */
-    public function edit()
+    public function edit()/*:void*/
     {
         $form = $this->getForm();
 
@@ -123,7 +121,7 @@ class il__PLUGIN_NAME__PluginGUI extends ilPageComponentPluginGUI
     /**
      *
      */
-    public function update()
+    public function update()/*:void*/
     {
         $form = $this->getForm();
 
@@ -148,20 +146,16 @@ class il__PLUGIN_NAME__PluginGUI extends ilPageComponentPluginGUI
     /**
      *
      */
-    public function cancel()
+    public function cancel()/*:void*/
     {
         $this->returnToParent();
     }
 
 
     /**
-     * @param string $a_mode
-     * @param array  $a_properties
-     * @param string $plugin_version
-     *
-     * @return string
+     * @inheritDoc
      */
-    public function getElementHTML($a_mode, array $a_properties, $plugin_version) : string
+    public function getElementHTML(/*string*/ $a_mode, array $a_properties, /*string*/ $plugin_version) : string
     {
         return ""; // TODO: Implement getElementHTML
     }