+++ templates/slots/Services/Repository/RepositoryObject/classes/class.ilObj__PLUGIN_NAME__Access.php
@@ -1,6 +1,6 @@
 <?php
 
-use __NAMESPACE__\Object\Obj;
+use __NAMESPACE__\ObjectSettings\ObjectSettings;
 use srag\DIC\__PLUGIN_NAME__\DICTrait;
 
 /**
@@ -42,21 +42,10 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
 
 
     /**
-     * @param string   $a_cmd
-     * @param string   $a_permission
-     * @param int|null $a_ref_id
-     * @param int|null $a_obj_id
-     * @param int|null $a_user_id
-     *
-     * @return bool
+     * @inheritDoc
      */
-    public function _checkAccess(/*string*/
-        $a_cmd, /*string*/
-        $a_permission, /*?int*/
-        $a_ref_id = null, /*?int*/
-        $a_obj_id = null, /*?int*/
-        $a_user_id = null
-    ) : bool {
+    public function _checkAccess(/*string*/ $a_cmd, /*string*/ $a_permission, /*?int*/ $a_ref_id = null, /*?int*/ $a_obj_id = null, /*?int*/ $a_user_id = null) : bool
+    {
         if ($a_ref_id === null) {
             $a_ref_id = filter_input(INPUT_GET, "ref_id");
         }
@@ -96,13 +85,8 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
      *
      * @return bool
      */
-    protected static function checkAccess(
-        string $a_cmd,
-        string $a_permission, /*?int*/
-        $a_ref_id = null, /*?int*/
-        $a_obj_id = null, /*?int*/
-        $a_user_id = null
-    ) {
+    protected static function checkAccess(string $a_cmd, string $a_permission, /*?int*/ $a_ref_id = null, /*?int*/ $a_obj_id = null, /*?int*/ $a_user_id = null) : bool
+    {
         return self::getInstance()->_checkAccess($a_cmd, $a_permission, $a_ref_id, $a_obj_id, $a_user_id);
     }
 
@@ -111,7 +95,7 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
      * @param object|string $class
      * @param string        $cmd
      */
-    public static function redirectNonAccess($class, string $cmd = "")
+    public static function redirectNonAccess($class, string $cmd = "")/*:void*/
     {
         ilUtil::sendFailure(self::plugin()->translate("permission_denied", ilObj__PLUGIN_NAME__GUI::LANG_MODULE_OBJECT), true);
 
@@ -126,17 +110,14 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
 
 
     /**
-     * @param int $obj_id
-     *
-     * @return bool
+     * @inheritDoc
      */
-    public static function _isOffline(/*?int*/
-        $a_obj_id
-    ) : bool {
-        $object = Obj::getObjectById(intval($a_obj_id));
+    public static function _isOffline(/*?int*/ $a_obj_id) : bool
+    {
+        $object_settings = ObjectSettings::getObjectSettingsById(intval($a_obj_id));
 
-        if ($object !== null) {
-            return (!$object->isOnline());
+        if ($object_settings !== null) {
+            return (!$object_settings->isOnline());
         } else {
             return true;
         }
@@ -148,9 +129,8 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
      *
      * @return bool
      */
-    public static function hasVisibleAccess(/*?int*/
-        $ref_id = null
-    ) : bool {
+    public static function hasVisibleAccess(/*?int*/ $ref_id = null) : bool
+    {
         return self::checkAccess("visible", "visible", $ref_id);
     }
 
@@ -160,9 +140,8 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
      *
      * @return bool
      */
-    public static function hasReadAccess(/*?int*/
-        $ref_id = null
-    ) : bool {
+    public static function hasReadAccess(/*?int*/ $ref_id = null) : bool
+    {
         return self::checkAccess("read", "read", $ref_id);
     }
 
@@ -172,9 +151,8 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
      *
      * @return bool
      */
-    public static function hasWriteAccess(/*?int*/
-        $ref_id = null
-    ) : bool {
+    public static function hasWriteAccess(/*?int*/ $ref_id = null) : bool
+    {
         return self::checkAccess("write", "write", $ref_id);
     }
 
@@ -184,9 +162,8 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
      *
      * @return bool
      */
-    public static function hasDeleteAccess(/*?int*/
-        $ref_id = null
-    ) : bool {
+    public static function hasDeleteAccess(/*?int*/ $ref_id = null) : bool
+    {
         return self::checkAccess("delete", "delete", $ref_id);
     }
 
@@ -196,9 +173,8 @@ class ilObj__PLUGIN_NAME__Access extends ilObjectPluginAccess
      *
      * @return bool
      */
-    public static function hasEditPermissionAccess(/*?int*/
-        $ref_id = null
-    ) : bool {
+    public static function hasEditPermissionAccess(/*?int*/ $ref_id = null) : bool
+    {
         return self::checkAccess("edit_permission", "edit_permission", $ref_id);
     }
 }