+++ templates/slots/Services/Repository/RepositoryObject/src/ObjectSettings/ObjectSettingsFormGUI.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace __NAMESPACE__\Object;
+namespace __NAMESPACE__\ObjectSettings;
 
 use il__PLUGIN_NAME__Plugin;
 use ilCheckboxInputGUI;
+use ilObj__PLUGIN_NAME__;
 use ilObj__PLUGIN_NAME__GUI;
 use ilTextAreaInputGUI;
 use ilTextInputGUI;
-use srag\CustomInputGUIs\__PLUGIN_NAME__\PropertyFormGUI\ObjectPropertyFormGUI;
+use srag\CustomInputGUIs\__PLUGIN_NAME__\PropertyFormGUI\ObjectSettingsPropertyFormGUI;
 
 /**
- * Class ObjSettingsFormGUI__VERSION_COMMENT__
+ * Class ObjectSettingsFormGUI__VERSION_COMMENT__
  *
- * @package __NAMESPACE__\Object
+ * @package __NAMESPACE__\ObjectSettings
  *
- * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
- * @author  __RESPONSIBLE_NAME__ <__RESPONSIBLE_EMAIL__>
+ * __AUTHOR_COMMENT__
  */
-class ObjSettingsFormGUI extends ObjectPropertyFormGUI
+class ObjectSettingsFormGUI extends ObjectPropertyFormGUI
 {
 
     const PLUGIN_CLASS_NAME = il__PLUGIN_NAME__Plugin::class;
@@ -25,11 +25,22 @@ class ObjSettingsFormGUI extends ObjectPropertyFormGUI
 
 
     /**
-     * @inheritdoc
+     * ObjectSettingsFormGUI constructor
+     *
+     * @param ilObj__PLUGIN_NAME__GUI $parent
+     * @param ilObj__PLUGIN_NAME__    $object
      */
-    protected function getValue(/*string*/
-        $key
-    ) {
+    public function __construct(ilObj__PLUGIN_NAME__GUI $parent, ilObj__PLUGIN_NAME__ $object)
+    {
+        parent::__construct($parent, $object);
+    }
+
+
+    /**
+     * @inheritDoc
+     */
+    protected function getValue(/*string*/ $key)
+    {
         switch ($key) {
             case "description":
                 return $this->object->getLongDescription();
@@ -41,7 +52,7 @@ class ObjSettingsFormGUI extends ObjectPropertyFormGUI
 
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      */
     protected function initCommands()/*: void*/
     {
@@ -52,7 +63,7 @@ class ObjSettingsFormGUI extends ObjectPropertyFormGUI
 
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      */
     protected function initFields()/*: void*/
     {
@@ -73,7 +84,7 @@ class ObjSettingsFormGUI extends ObjectPropertyFormGUI
 
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      */
     protected function initId()/*: void*/
     {
@@ -82,7 +93,7 @@ class ObjSettingsFormGUI extends ObjectPropertyFormGUI
 
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      */
     protected function initTitle()/*: void*/
     {