+++ templates/slots/Services/UIComponent/UserInterfaceHook/classes/class.il__PLUGIN_NAME__Plugin.php
@@ -47,7 +47,7 @@ class il__PLUGIN_NAME__Plugin extends ilUserInterfaceHookPlugin
 
 
     /**
-     * @return string
+     * @inheritDoc
      */
     public function getPluginName() : string
     {
@@ -56,9 +56,9 @@ class il__PLUGIN_NAME__Plugin extends ilUserInterfaceHookPlugin
 
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      */
-    public function updateLanguages($a_lang_keys = null)
+    public function updateLanguages(/*?array*/ $a_lang_keys = null)/*:void*/
     {
         parent::updateLanguages($a_lang_keys);
 
@@ -68,7 +68,7 @@ class il__PLUGIN_NAME__Plugin extends ilUserInterfaceHookPlugin
 
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      */
     protected function deleteData()/*: void*/
     {