+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.2.0]
+- Rename `Obj` to `ObjectSettings`
+- Fix author comment
+- Other improvments
+
 ## [1.1.0]
 - Adjustments for latest RemovePluginDataConfirm library